+++ meta/block_pass.rb
@@ -5,4 +5,21 @@ Mutant::Meta::Example.add :block_pass do
 
   singleton_mutations
   mutation 'foo'
+  mutation 'foo(&nil)'
+  mutation 'foo(&self)'
+end
+
+Mutant::Meta::Example.add :block_pass do
+  source 'foo(&method(:bar))'
+
+  singleton_mutations
+  mutation 'foo'
+  mutation 'foo(&nil)'
+  mutation 'foo(&self)'
+  mutation 'foo(&method)'
+  mutation 'foo(&method(nil))'
+  mutation 'foo(&method(self))'
+  mutation 'foo(&method(:bar__mutant__))'
+  mutation 'foo(&public_method(:bar))'
+  mutation 'foo(&:bar)'
 end